Refactor provisioners, and more
Any provisioner appended with 'local' can only
be used if the shell-local provisioner of packer is used.

Added guest-additions provisioner, that way templates
can in addition call this provisioner instead of having
each template have their own version of the code.

Finally, as part of the naming schema for provisioners
and processors. If any particular can run on any Linux OS,
then really the lowest common dependency here
is the shell that the provisioner/processor runs in.
Which in most (if not all cases) will be bash.

shell-provisioners env file is now a config file. Really
this makes more sense then thinking of the file as a env var
when really its for configuration.